fix theatre-square: int->int64

+++ theatre-square/main.go
@@ -5,7 +5,7 @@ import (
 )
 
 func main() {
-	var n, m, a int
+	var n, m, a int64
 	fmt.Scan(&n, &m, &a)
 
 	fmt.Println(((n + a - 1) / a) * ((m + a - 1) / a))
